requirements.txt
django 
django-rest-framework
django-cors-headers
opencv-python
numpy
Pillow
deepface
